momentum_strategy_project/requirements.txt
# Core dependencies
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0

# Data interfaces
wrds>=3.1.0
requests>=2.31.0
pyyaml>=6.0

# Analysis
statsmodels>=0.14.0
scikit-learn>=1.3.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Utilities
loguru>=0.7.0
tqdm>=4.65.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Code quality
flake8>=6.0.0
mypy>=1.4.0
black>=23.0.0
